split up table into multiple files
